Moved ginga to plugins/ginga_viewer
