+++ glue/plugins/ginga_viewer/client.py
@@ -5,13 +5,13 @@ from time import time
 
 import numpy as np
 
-from ..core.exceptions import IncompatibleAttribute
-from ..core.util import Pointer, split_component_view
-from ..utils import view_shape, stack_view, color2rgb
+from ...core.exceptions import IncompatibleAttribute
+from ...core.util import Pointer, split_component_view
+from ...utils import view_shape, stack_view, color2rgb
 
-from .image_client import ImageClient
-from .layer_artist import (LayerArtistBase,
-                           ImageLayerBase, SubsetImageLayerBase)
+from ...clients.image_client import ImageClient
+from ...clients.layer_artist import (LayerArtistBase,
+                                    ImageLayerBase, SubsetImageLayerBase)
 
 from ginga.util import wcsmod
 from ginga.misc import Bunch
