+++ glue/plugins/ginga_viewer/tests/test_client.py
@@ -4,25 +4,23 @@ import pytest
 import numpy as np
 from numpy.testing import assert_allclose, assert_array_equal
 
-from ...core import Data
+from ginga.qtw.ImageViewCanvasQt import ImageViewCanvas
+from ginga.misc import log
 
-from .test_image_client import _TestImageClientBase
+from ....core import Data
 
-try:
-    from ..ginga_client import GingaClient, SubsetImage, BaseImage
-    from ginga.qtw.ImageViewCanvasQt import ImageViewCanvas
-    from ginga.misc import log
-except ImportError:
-    pass
+from ....clients.tests.test_image_client import _TestImageClientBase
 
-from ...tests.helpers import requires_ginga
+from ..client import GingaClient, SubsetImage, BaseImage
+
+from ....tests.helpers import requires_ginga
 
 
 @requires_ginga
 class TestGingaClient(_TestImageClientBase):
 
     def new_client(self, dc=None, canvas=None):
-        from ...qt import get_qapp
+        from ....qt import get_qapp
         get_qapp()
         dc = dc or self.collect
         l = log.get_logger(name='ginga', log_stderr=True)
